fix bug Rain.java int -> double

+++ app/src/main/java/com/example/myweather/data/pojo/day/Rain.java
@@ -6,13 +6,13 @@ import com.google.gson.annotations.SerializedName;
 public class Rain {
     @SerializedName("1h")
     @Expose
-    private int _1h;
+    private double _1h;
 
-    public int get1h() {
+    public double get1h() {
         return _1h;
     }
 
-    public void set1h(int _1h) {
+    public void set1h(double _1h) {
         this._1h = _1h;
     }
 }
